backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==23.5
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
psycopg[binary]==3.1.19
Pillow==10.2.0
WeasyPrint==60.2
python-decouple==3.8
whitenoise==6.6.0
openpyxl==3.1.2
reportlab>=4.0
python-bidi>=0.6
drf-spectacular==0.27.1
chardet>=5.0
gunicorn>=21.2.0
qrcode[pil]>=7.4
stripe>=14.0
